Moves DPLA Api key to a constant defined in the new config.php file
This means a new step when setting up a new site: entering the key in
the config file. However, this also means that we won't be publishing
our key on GitHub (as long as we don't accidentally push it up from the
config file!). It also has the bonus of easier maintenance if we need to
change the key for some reason, which is related to issue referenced
below: the key was no longer valid.

part of https://github.com/NEU-DSG/toolkit_requests/issues/117